Split OGTagPreviewPage into several files for easier grocking
